pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "insurance-recommendation-agent"
version = "0.1.0"
description = "Insurance recommendation agent using Google ADK"
requires-python = ">=3.12"
dependencies = [
  "google-adk>=1.31.1",
  "google-genai>=1.73.1",
  "google-cloud-aiplatform>=1.148.1",
  "google-cloud-spanner>=3.65.0",
  "google-cloud-storage>=3.10.1",
  "google-cloud-secret-manager>=2.27.0",
  "google-cloud-logging>=3.15.0",
  "google-cloud-discoveryengine>=0.13.12",
  "google-cloud-speech>=2.38.0",
  "google-cloud-trace>=1.19.0",
  "google-cloud-monitoring>=2.30.0",
  "google-cloud-pubsub>=2.37.0",
  "google-cloud-bigquery>=3.41.0",
  "google-cloud-bigtable>=2.36.0",
  "google-cloud-dataplex>=2.18.0",
  "google-cloud-iam>=2.22.0",
  "google-cloud-resource-manager>=1.17.0",
  "fastapi>=0.136.0",
  "uvicorn>=0.44.0",
  "python-dotenv>=1.2.2",
  "pydantic>=2.13.3",
  "pydantic-settings>=2.14.0",
  "sqlalchemy[asyncio]>=2.0.49",
  "sqlalchemy-spanner>=1.17.3",
  "aiosqlite>=0.22.1",
  "alembic>=1.18.4",
  "authlib>=1.7.0",
  "graphviz>=0.21",
  "mcp>=1.27.0",
  "requests>=2.33.1",
  "toolbox-adk>=1.0.0",
  "toolbox-core>=1.0.0",
  "greenlet>=3.4.0",
  "pytest>=8.4.2",
]

[project.optional-dependencies]
eval = ["google-adk[eval]>=1.31.1"]

[tool.setuptools.packages.find]
include = ["app*"]
